requirements.txt
asyncpg>=0.30.0
fastapi>=0.110.0
greenlet>=3.0.3
httpx>=0.28.1
psycopg[binary]>=3.2.10
pydantic-settings>=2.11.0
pyyaml>=6.0.3
sqlalchemy>=2.0.44
uvicorn[standard]>=0.30.0
yfinance>=0.2.38
vaderSentiment>=3.3.2
# LLM clients
openai>=1.0.0
anthropic>=0.30.0
google-generativeai>=0.5.0
tenacity>=8.0.0
# Report generation
jinja2>=3.1.0
weasyprint>=60.0
markdown>=3.5.0
aiosmtplib>=3.0.0
